scrapper/requirements.txt
selenium
bs4
webdriver-manager
kafka-python
